pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/microsoftdesigner"]

[project]
name = "microsoftdesigner"
version = "0.0.12"
description = "High quality image generation by Microsoft Designer. Reverse engineered API."
authors = [
    {name = "imnotdev25",email = "85677268+imnotdev25@users.noreply.github.com"}
]
license = {text = "GNU General Public License v3.0"}
license-files = ["LICENSE"]
readme = "README.md"
packages = [
    {include = "src/microsoftdesigner"}
]
requires-python = ">=3.8"
keywords = ["msdesigner", "microsoft", "designer", "api"]
dependencies = [
    "httpx>=0.28.1",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3",
    "python-dotenv>=0.21.1",
]

[project.urls]
Homepage = "https://github.com/imnotdev25/Msdesigner-api"
Issues = "https://github.com/imnotdev25/Msdesigner-api/issues"
